requirements.txt
python-telegram-bot[job-queue]>=20.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
textblob>=0.17.1
feedparser>=6.0.10
playwright>=1.45.0
